requirements.txt
beautifulsoup4==4.10.0
markdown2==2.4.1
openai==0.27.0
Pillow==8.4.0
PyMuPDF==1.18.19
python-dotenv==0.19.2
streamlit==0.88.0